competences

+++ src/main/java/edu/isep/beans/SousCompetences.java
@@ -3,7 +3,7 @@ package edu.isep.beans;
 public class SousCompetences {
 	private int id;
 	private String sous_competences;
-	private double competences_id;
+	private int competences_id;
 	public int getId() {
 		return id;
 	}
@@ -16,12 +16,14 @@ public class SousCompetences {
 	public void setSous_competences(String sous_competences) {
 		this.sous_competences = sous_competences;
 	}
-	public double getCompetences_id() {
+	public int getCompetences_id() {
 		return competences_id;
 	}
-	public void setCompetences_id(double competences_id) {
+	public void setCompetences_id(int competences_id) {
 		this.competences_id = competences_id;
 	}
+
+
